data-generation/requirements.txt
kafka-python==2.0.2
faker==19.12.0
numpy==1.24.3
pandas==2.0.3
python-dateutil==2.8.2
geopy==2.4.0
pydantic==2.4.2
psycopg2-binary==2.9.9
